Update sport challenge goal value type

+++ app/src/main/java/gr/dipae/thesisfitnessapp/data/user/challenges/model/RemoteSportChallenges.kt
@@ -13,5 +13,5 @@ data class RemoteSportChallenges(
 
 data class RemoteChallengeGoal(
     val type: String = "",
-    val value: Double = 0.0
+    val value: Long = 0
 )
